COZMO-13964 escapig special characters for JP specifically (#5371)

+++ tools/smartling/strip-zero-width-spaces.py
@@ -16,14 +16,16 @@ Logger.addHandler(stdout_handler)
 
 def strip_zero_width_spaces(args):
   for root, dirs, files in os.walk(args.localized_strings_dir):
-    for file in files:
-      if file.endswith('.json'):
-        with open(os.path.join(root, os.path.basename(file)), "r+", encoding='utf8') as inoutfile:
-          lines = [line.replace(u'\u200b', '') for line in inoutfile]   # ZERO-WIDTH SPACE
-          lines = [line.replace(u'\u00A0', '') for line in lines]       # NO-BREAK SPACE
-          inoutfile.seek(0)
-          inoutfile.truncate()
-          inoutfile.writelines(lines)
+    for dir in dirs:
+      if dir == 'ja-JP':
+        for file in files:
+          if file.endswith('.json'):
+            with open(os.path.join(root, os.path.basename(file)), "r+", encoding='utf8') as inoutfile:
+              lines = [line.replace(u'\u200b', '') for line in inoutfile]   # ZERO-WIDTH SPACE
+              lines = [line.replace(u'\u00A0', '') for line in lines]       # NO-BREAK SPACE
+              inoutfile.seek(0)
+              inoutfile.truncate()
+              inoutfile.writelines(lines)
 
 ####################################
 ############### MAIN ###############
